+++ examples/06-body-parser/src/app/modules/routed/some/some.module.ts
@@ -2,10 +2,11 @@ import { featureModule } from '@ditsmod/core';
 import { RoutingModule } from '@ditsmod/routing';
 
 import { SomeController } from './some.controller.js';
+import { SingletonController } from './singleton-some.controller.js';
 
 @featureModule({
   imports: [RoutingModule],
-  controllers: [SomeController],
+  controllers: [SomeController, SingletonController],
 })
 export class SomeModule {
 }