+++ examples/06-body-parser/test/main.spec.ts
@@ -29,4 +29,19 @@ describe('06-body-parser', () => {
       .expect(200)
       .expect({ one: 1 });
   });
+
+  it('controller singleton should works with get', async () => {
+    await request(server)
+      .get('/singleton')
+      .expect(200)
+      .expect('Hello, you need send POST request');
+  });
+
+  it('controller singleton should parsed post', async () => {
+    await request(server)
+      .post('/singleton')
+      .send({ one: 1 })
+      .expect(200)
+      .expect({ one: 1 });
+  });
 });