+++ packages/body-parser/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="body-parser-2.13.1"></a>
+# [body-parser-2.13.1](https://github.com/ditsmod/ditsmod/releases/tag/body-parser-2.13.1) (2023-09-30)
+
+### Bug fixes
+
+- Added support for controller singleton.
+
 <a name="body-parser-2.13.0"></a>
-# [body-parser-2.13.0](https://github.com/ditsmod/ditsmod/releases/tag/body-parser-2.13.0) (2023-09-30)
+## [body-parser-2.13.0](https://github.com/ditsmod/ditsmod/releases/tag/body-parser-2.13.0) (2023-09-30)
 
 ### Features